# orbita/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: |
    API specification for the internal ÓRBITA discovery workspace.
    All operations except health require an authenticated Clerk session cookie.
    Workspace roles are assigned in trusted Clerk public metadata under
    `workspaceRole`: `read_only`, `contributor`, `facilitator`, `editor`, or
    `administrator`. Unknown or missing roles fail closed to `read_only`.
servers:
  - url: /api
    description: Base API path
...
security:
  - clerkSession: []
paths:
...
  /users:
    get:
      operationId: getWorkspaceUsers
      tags: [access]
      summary: List workspace users and their roles
      description: Administrator access required.
      responses:
        "200":
          description: Workspace users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkspaceUser"
        "403":
          description: Administrator role required
  /users/{id}/role:
    patch:
      operationId: updateWorkspaceUserRole
      tags: [access]
      summary: Change a workspace user's role
      description: Administrator access required. The final administrator cannot be demoted.
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceRoleUpdate"
      responses:
        "200":
          description: Updated workspace user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceUser"
        "400":
          description: Invalid role update
        "403":
          description: Administrator role required
        "404":
          description: User not found
        "409":
          description: The final administrator cannot be demoted
...
components:
  securitySchemes:
    clerkSession:
      type: apiKey
      in: cookie
      name: __session
      description: Clerk browser session cookie for an individually signed-in team member.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    WorkspaceRole:
      type: string
      enum: [read_only, contributor, facilitator, editor, administrator]
...
    WorkspaceUser:
      type: object
      required: [id, name, email, emailVerified, role, status, createdAt, lastActiveAt, isCurrentUser]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        emailVerified: { type: boolean }
        role:
          $ref: "#/components/schemas/WorkspaceRole"
        status: { type: string, enum: [active, locked, banned] }
        createdAt: { type: string, format: date-time }
        lastActiveAt: { type: ["string", "null"], format: date-time }
        isCurrentUser: { type: boolean }
    WorkspaceRoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          $ref: "#/components/schemas/WorkspaceRole"
...