# orbita/lib/api-zod/src/generated/types/workspaceRole.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for the internal ÓRBITA discovery workspace.
 * All operations except health require an authenticated Clerk session cookie.
 * Workspace roles are assigned in trusted Clerk public metadata under
 * `workspaceRole`: `read_only`, `contributor`, `facilitator`, `editor`, or
 * `administrator`. Unknown or missing roles fail closed to `read_only`.
 *
 * OpenAPI spec version: 0.1.0
 */

export type WorkspaceRole = typeof WorkspaceRole[keyof typeof WorkspaceRole];


export const WorkspaceRole = {
  read_only: 'read_only',
  contributor: 'contributor',
  facilitator: 'facilitator',
  editor: 'editor',
  administrator: 'administrator',
} as const;